third_party/blink/web_tests/wpt_internal/display-lock/css-render-subtree/parsing/render-subtree-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS RenderSubtree Test: render-subtree with valid values</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<meta name="assert" content="render-subtree supports the full grammar 'none | [invisible || skip-activation || skip-viewport-activation]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("render-subtree", "none");
test_valid_value("render-subtree", "invisible");
test_valid_value("render-subtree", "skip-activation");
test_valid_value("render-subtree", "skip-viewport-activation");
test_valid_value("render-subtree", "invisible skip-activation");
test_valid_value("render-subtree", "invisible skip-viewport-activation");
test_valid_value("render-subtree", "skip-activation skip-viewport-activation");
test_valid_value("render-subtree", "invisible skip-activation skip-viewport-activation");
test_valid_value("render-subtree", "skip-activation skip-viewport-activation invisible");
</script>
</body>
</html>
